refactored routes, and removed credit card / merchants
