+++ server/api/index.js
@@ -1,16 +1,14 @@
 const router = require('express').Router()
 module.exports = router
 
-router.use('/brands', require('./brandRoutes'))
-router.use('/creditCards', require('./creditCardRoutes'))
-router.use('/departments', require('./departmentRoutes'))
-router.use('/merchants', require('./merchantRoutes'))
-router.use('/products', require('./productsRoute'))
-router.use('/reviews', require('./reviewRoutes'))
-router.use('/users', require('./usersRoutes'))
-router.use('/userOrders', require('./userOrderRoutes'))
-router.use('/orderProducts', require('./orderProductRoutes'))
-router.use('/images', require('./imageRoute'))
+router.use('/brands', require('./routes/brand'))
+router.use('/departments', require('./routes/department'))
+router.use('/products', require('./routes/product'))
+router.use('/reviews', require('./routes/review'))
+router.use('/users', require('./routes/user'))
+router.use('/userOrders', require('./routes/userOrder'))
+router.use('/orderProducts', require('./routes/orderProduct'))
+router.use('/images', require('./routes/image'))
 
 router.use((req, res, next) => {
   const error = new Error('Not Found')
