+++ server/api/routes/department.js
@@ -1,13 +1,13 @@
-const {Brand, Product, Review, Image} = require('../db/models')
 const router = require('express').Router()
+const {Brand, Department, Product, Review, Image} = require('../../db/models')
 const Sequelize = require('sequelize')
 
 module.exports = router
 
 router.get('/', async (req, res, next) => {
   try {
-    const brand = await Brand.findAll()
-    res.status(200).send(brand)
+    const department = await Department.findAll()
+    res.status(200).send(department)
   } catch (err) {
     next(err)
   }
@@ -16,7 +16,7 @@ router.get('/', async (req, res, next) => {
 router.get('/:id', async (req, res, next) => {
   const id = req.params.id
   try {
-    const brand = await Brand.findByPk(id, {
+    const department = await Department.findByPk(id, {
       include: [
         {
           model: Product,
@@ -42,13 +42,13 @@ router.get('/:id', async (req, res, next) => {
         }
       ],
       group: [
-        'brand.id',
+        'department.id',
         'products.id',
         'products->brand.id',
         'products->images.id'
       ]
     })
-    res.status(200).send(brand)
+    res.status(200).send(department)
   } catch (err) {
     next(err)
   }
@@ -56,8 +56,8 @@ router.get('/:id', async (req, res, next) => {
 
 router.post('/', async (req, res, next) => {
   try {
-    const brand = await Brand.create(req.body)
-    res.status(201).send(brand)
+    const department = await Department.create(req.body)
+    res.status(201).send(department)
   } catch (err) {
     next(err)
   }
@@ -66,9 +66,9 @@ router.post('/', async (req, res, next) => {
 router.put('/:id', async (req, res, next) => {
   const id = req.params.id
   try {
-    const brand = await Brand.findByPk(id)
-    brand.update(req.body)
-    res.status(200).send(brand)
+    const department = await Department.findByPk(id)
+    department.update(req.body)
+    res.status(200).send(department)
   } catch (err) {
     next(err)
   }
@@ -77,8 +77,8 @@ router.put('/:id', async (req, res, next) => {
 router.delete('/:id', async (req, res, next) => {
   const id = req.params.id
   try {
-    const brand = await Brand.findByPk(id)
-    brand.destroy()
+    const department = await Department.findByPk(id)
+    department.destroy()
     res.sendStatus(204)
   } catch (err) {
     next(err)
