+++ server/api/routes/review.js
@@ -1,5 +1,5 @@
 const router = require('express').Router()
-const {Review} = require('../db/models')
+const {Review} = require('../../db/models')
 module.exports = router
 
 router.get('/', async (req, res, next) => {
