+++ server/api/routes/user.js
@@ -6,7 +6,7 @@ const {
   OrderProduct,
   Product,
   Brand
-} = require('../db/models')
+} = require('../../db/models')
 module.exports = router
 
 router.get('/', async (req, res, next) => {
